[UI][Enemy][Mana]

+++ Assets/Scripts/Spells/Potions/Potion.cs
@@ -6,7 +6,7 @@ public class Potion : ScriptableObject
     public string Name;
     public Sprite icon;
     public PotionEffect potionEffect;
-    public float effectStrength; 
+    public int effectStrength; 
     public int  duration; //in seconds - 0 if instant
 
 }
